refactor(data): rename implementations module to `Impls`
