+++ src/pamiq_core/data/__init__.py
@@ -1,6 +1,6 @@
 from .buffer import BufferData, DataBuffer, StepData
 from .container import DataCollectorsDict, DataUsersDict
-from .implementations import RandomReplacementBuffer, SequentialBuffer
+from .impls import RandomReplacementBuffer, SequentialBuffer
 from .interface import DataCollector, DataUser
 
 __all__ = [
