+++ src/pamiq_core/data/impls/sequential_buffer.py
